+++ src/RavenNest.DataModels/Clan.cs
@@ -7,7 +7,7 @@ namespace RavenNest.DataModels
         private Guid id; public Guid Id { get => id; set => Set(ref id, value); }
         private Guid userId; public Guid UserId { get => userId; set => Set(ref userId, value); }
         private int level; public int Level { get => level; set => Set(ref level, value); }
-        private decimal experience; public decimal Experience { get => experience; set => Set(ref experience, value); }
+        private double experience; public double Experience { get => experience; set => Set(ref experience, value); }
         private string name; public string Name { get => name; set => Set(ref name, value); }
         private string logo; public string Logo { get => logo; set => Set(ref logo, value); }
         private int nameChangeCount; public int NameChangeCount { get => nameChangeCount; set => Set(ref nameChangeCount, value); }
