+++ src/RavenNest.Models/PlayerState.cs
@@ -8,8 +8,8 @@ namespace RavenNest.Models
         public Guid CharacterId { get; set; }
         public string CurrentTask { get; set; }
         public int[] Level { get; set; }
-        public decimal[] Experience { get; set; }
-        public decimal[] Statistics { get; set; }
+        public double[] Experience { get; set; }
+        public double[] Statistics { get; set; }
         public float SyncTime { get; set; }
         public int Revision { get; set; }
     }
