+++ src/RavenNest.Models/Resources.cs
@@ -5,13 +5,13 @@ namespace RavenNest.Models
     public class Resources
     {
         public Guid Id { get; set; }
-        public decimal Wood { get; set; }
-        public decimal Ore { get; set; }
-        public decimal Fish { get; set; }
-        public decimal Wheat { get; set; }
-        public decimal Magic { get; set; }
-        public decimal Arrows { get; set; }
-        public decimal Coins { get; set; }
+        public double Wood { get; set; }
+        public double Ore { get; set; }
+        public double Fish { get; set; }
+        public double Wheat { get; set; }
+        public double Magic { get; set; }
+        public double Arrows { get; set; }
+        public double Coins { get; set; }
         public int Revision { get; set; }
     }
 }
