+++ src/RavenNest.Models/ScrollInfo.cs
@@ -5,10 +5,10 @@ namespace RavenNest.Models
     public class ScrollInfo
     {
         public Guid ItemId { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string Name { get; set; }
         public ScrollInfo() { }
-        public ScrollInfo(Guid itemId, string name, decimal amount)
+        public ScrollInfo(Guid itemId, string name, double amount)
         {
             ItemId = itemId;
             Name = name;
